+++ Joufflu.Samples/Views/Inputs/DropdownSamples.xaml.cs
@@ -1,6 +1,6 @@
 using System.Windows.Controls;
 
-namespace Joufflu.Samples.Views.Navigation;
+namespace Joufflu.Samples.Views.Inputs;
 
 public partial class DropdownSamples : UserControl
 {